Move config to internal
