+++ internal/config/types_azure.go
