+++ internal/config/types_global.go
