+++ internal/generator/filters.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/flosch/pongo2/v5"
-	"github.com/labd/mach-composer/config"
+	"github.com/labd/mach-composer/internal/config"
 	"github.com/sirupsen/logrus"
 )
