+++ internal/generator/writer_test.go
@@ -3,7 +3,7 @@ package generator
 import (
 	"testing"
 
-	"github.com/labd/mach-composer/config"
+	"github.com/labd/mach-composer/internal/config"
 	"github.com/stretchr/testify/assert"
 )
