+++ internal/updater/changes.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/labd/mach-composer/config"
+	"github.com/labd/mach-composer/internal/config"
 )
 
 type UpdateSet struct {
